more windows fixes

+++ frida-gum/src/range_details.rs
@@ -15,17 +15,17 @@ use crate::MemoryRange;
 #[derive(FromPrimitive)]
 #[repr(u32)]
 pub enum PageProtection {
-    NoAccess = gum_sys::_GumPageProtection_GUM_PAGE_NO_ACCESS,
-    Read = gum_sys::_GumPageProtection_GUM_PAGE_READ,
-    Write = gum_sys::_GumPageProtection_GUM_PAGE_WRITE,
-    Execute = gum_sys::_GumPageProtection_GUM_PAGE_EXECUTE,
-    ReadWrite =
-        gum_sys::_GumPageProtection_GUM_PAGE_READ | gum_sys::_GumPageProtection_GUM_PAGE_WRITE,
-    ReadExecute =
-        gum_sys::_GumPageProtection_GUM_PAGE_READ | gum_sys::_GumPageProtection_GUM_PAGE_EXECUTE,
-    ReadWriteExecute = gum_sys::_GumPageProtection_GUM_PAGE_READ
-        | gum_sys::_GumPageProtection_GUM_PAGE_WRITE
-        | gum_sys::_GumPageProtection_GUM_PAGE_EXECUTE,
+    NoAccess = gum_sys::_GumPageProtection_GUM_PAGE_NO_ACCESS as u32,
+    Read = gum_sys::_GumPageProtection_GUM_PAGE_READ as u32,
+    Write = gum_sys::_GumPageProtection_GUM_PAGE_WRITE as u32,
+    Execute = gum_sys::_GumPageProtection_GUM_PAGE_EXECUTE as u32,
+    ReadWrite = gum_sys::_GumPageProtection_GUM_PAGE_READ as u32
+        | gum_sys::_GumPageProtection_GUM_PAGE_WRITE as u32,
+    ReadExecute = gum_sys::_GumPageProtection_GUM_PAGE_READ as u32
+        | gum_sys::_GumPageProtection_GUM_PAGE_EXECUTE as u32,
+    ReadWriteExecute = gum_sys::_GumPageProtection_GUM_PAGE_READ as u32
+        | gum_sys::_GumPageProtection_GUM_PAGE_WRITE as u32
+        | gum_sys::_GumPageProtection_GUM_PAGE_EXECUTE as u32,
 }
 
 /// The file association to a page.
