+++ frida-gum/src/stalker/event_sink.rs
@@ -13,12 +13,12 @@ use std::os::raw::c_void;
 #[repr(u32)]
 #[cfg_attr(doc_cfg, doc(cfg(feature = "event-sink")))]
 pub enum EventMask {
-    None = gum_sys::_GumEventType_GUM_NOTHING,
-    Call = gum_sys::_GumEventType_GUM_CALL,
-    Ret = gum_sys::_GumEventType_GUM_RET,
-    Exec = gum_sys::_GumEventType_GUM_EXEC,
-    Block = gum_sys::_GumEventType_GUM_BLOCK,
-    Compile = gum_sys::_GumEventType_GUM_COMPILE,
+    None = gum_sys::_GumEventType_GUM_NOTHING as u32,
+    Call = gum_sys::_GumEventType_GUM_CALL as u32,
+    Ret = gum_sys::_GumEventType_GUM_RET as u32,
+    Exec = gum_sys::_GumEventType_GUM_EXEC as u32,
+    Block = gum_sys::_GumEventType_GUM_BLOCK as u32,
+    Compile = gum_sys::_GumEventType_GUM_COMPILE as u32,
 }
 
 #[cfg_attr(doc_cfg, doc(cfg(feature = "event-sink")))]
